Move test outside of pages directory
